use the real ioctl on a fake file to test ioctl error
this proves that the error response from the real ioctl (-1) is the error
response we are looking out for

will squash in some more changes related to deleting stuff we don't use
anymore now that I fixed the coverage report (mainly in a previous commit
by changing the bytes program - which ironically we don't really need
now that we are testing the real ioctl here)

partially Revert "test errors with files and ioctl and stuff"

This partially reverts commit ad48e668800f52c7e27481590f431a378347168c.

remove the intentional failure from fake ioctl - now that we create
intentional failures by using real ioctl and trying it out on the fake
uinput

intermediate files!

so the change in the previous commit - how bytes is compiled wasn't
directly related, but might have contributed

I was getting a message that said gcno timestamp mismatch or something
that led me to this file in gcc
https://github.com/gcc-mirror/gcc/blob/eb83605be3db9e8246c73755eafcac5df32ddc69/libgcc/libgcov-driver.c

the error was hard to find because in master they changed it from

"timestamp" to "checksum" - and I am still on gcc-11

the commit that changed it is
https://github.com/gcc-mirror/gcc/commit/72e0c742bd01f8e7e6dcca64042b9ad7e75979de

> gcov: make profile merging smarter
Support merging of profiles that are built from a different .o files
but belong to the same source file. Moreover, a checksum is verified
during profile merging and so we can safely combine such profile.

I tried installing and compiling with gcc-12 - but there was another error
but then I looked throught the .o files to see why I had more than one
anyhow, and I saw that some were missing
and I vaguely remembered seeing some rm commands that I didn't recognise
a long time ago
I looked through the log and saw that the .o files were being deleted
and then with --debug=b - it said it was deleting "intermediate files"
and then I found out on stackoverflow (link pasted) that that's what it
does - which seems to defeat the object of all that other caching I did
I guess they are worried about caching during the same run - but still
surprised

so when I added the .o files to the precious list everything worked as
hoped

can also delete lookup.lua

it was only there to look up the codes to cause all the fake failures
of ioctl

reminder the two things that made this unecessary were:
- using the real ioctl with a fake uinput to cause an error
- a single error handling "routine" - and not testing every possible error
(a few commits ago)

undo unrelated clean up about creating test fifo